[ucsim-i8048] Implement diffs between 802x,8048 and 8041
git-svn-id: http://svn.code.sourceforge.net/p/sdcc/code/trunk/sdcc@15405 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sim/ucsim/src/sims/pdk.src/osc.cc
@@ -67,7 +67,7 @@ cl_osc::init(void)
 	      cfg_help(osc_freq_eosc));
   cfg_cell(osc_freq_eosc)->decode(&fre);
 
-  fre= puc->get_xtal();
+  fre= (unsigned int)(puc->get_xtal());
   frh= 16000000;
   frl= 24000;
   reset();
